Update subscriptions api
