+++ src/App/app.module.ts
@@ -6,7 +6,6 @@ import { ApolloDriver, ApolloDriverConfig } from '@nestjs/apollo'
 import { GraphQLDirective, DirectiveLocation } from 'graphql'
 import { EventEmitterModule } from '@nestjs/event-emitter'
 import { SentryModule } from '@ntegral/nestjs-sentry'
-import { BullModule } from '@nestjs/bull'
 import WikiResolver from './wiki.resolver'
 import LanguageResolver from './language.resolver'
 import CategoryResolver from './category.resolver'
@@ -38,7 +37,6 @@ import WikiSubscriptionResolver from './subscriptions.resolver'
 import WikiSubscriptionService from './subscriptions.service'
 import TokenValidator from './utils/validateToken'
 import SentryPlugin from '../sentry/sentryPlugin'
-import NotificationsModule from './notifications/notifications.module'
 import PgNotificationsQueue from './notifications/pgQueue'
 
 @Module({
@@ -74,15 +72,6 @@ import PgNotificationsQueue from './notifications/pgQueue'
       }),
       inject: [ConfigService],
     }),
-    BullModule.forRootAsync({
-      useFactory: () => ({
-        redis: {
-          host: 'localhost',
-          port: 6379,
-        },
-      }),
-    }),
-    NotificationsModule,
     MailerModule,
     httpModule(20000),
     EventEmitterModule.forRoot(),
