+++ src/App/notifications/pgQueue.ts
@@ -1,10 +1,15 @@
 import { Injectable, Logger, OnModuleInit } from '@nestjs/common'
 import { ConfigService } from '@nestjs/config'
 import PgBoss from 'pg-boss'
+import { Connection } from 'typeorm'
+import Subscription from '../../Database/Entities/subscription.entity'
 
 @Injectable()
 export default class PgNotificationsQueue implements OnModuleInit {
-  constructor(private configService: ConfigService) {}
+  constructor(
+    private configService: ConfigService,
+    private connection: Connection,
+  ) {}
 
   public static readonly queueName = 'notifications'
 
@@ -18,5 +23,25 @@ export default class PgNotificationsQueue implements OnModuleInit {
     await boss.start()
 
     boss.on('error', error => Logger.error(error))
+    // await boss.subscribe(PgNotificationsQueue.queueName, 'any')
+    await boss.work(
+      PgNotificationsQueue.queueName,
+      this.handleWikiUpdate,
+    )
+  }
+
+  async handleWikiUpdate(job: any) {
+    const repository = this.connection.getRepository(Subscription)
+    const usersSubscribed = await repository.find({
+      auxiliaryId: job.data.id as string,
+    })
+
+    for await (const user of usersSubscribed) {
+      // TODO: send mails and push notifications here
+      console.log(user.email)
+    }
+    // console.log(usersSubscribed)
+    console.log(job.data)
+    return true
   }
 }
